requirements.txt
# Python 3.11.11
fastapi==0.115.6
uvicorn==0.34.0
jpype1==1.5.1
python-dotenv==1.0.1
pytest==8.3.4
pydantic==2.10.4
SQLAlchemy==2.0.36
ipykernel==6.29.5
psycopg2-binary==2.9.10
httpx==0.28.1
bcrypt==4.2.1
pytest-cov==6.0.0
setuptools